// frontend/src/app/core/models/SolicitudAcademica/DTOResponse/SolicitudAcademicaResumenDTORespuesta.ts
export interface SolicitudAcademicaResumenDTORespuesta {
  uuidSolicitudAcademica: string;
  radicado: string;
  uuidTipoSolicitudAcademica: string;
  tipoSolicitud: string;
  fechaCreacion: string;
  etiqueta: string;
  etapaCodigo: string;
  nombreEstudiante: string;
  codigoEstudiantil: string;
}